django_graphene/requirements.txt
aniso8601==3.0.2
bcrypt==3.1.6
cffi==1.12.2
Django==2.1.4
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.2.1
django-redis==4.10.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
pbr==5.1.1
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.1.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
stevedore==1.30.0
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
